# apps/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ErrorResponse:
    properties:
      message:
        example: not implemented
        type: string
    type: object
  handlers.LogoutRequest:
    properties:
      refreshToken:
        example: sample-refresh-token
        type: string
    type: object
  handlers.RefreshTokenRequest:
    properties:
      refreshToken:
        example: sample-refresh-token
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      displayName:
        example: 냉장고고수
        type: string
      signupToken:
        example: sample-signup-token
        type: string
    type: object
...
info:
  contact: {}
  description: API for naengrijeongdon inventory management.
  title: NaengriJeongdon API
  version: 0.1.0
paths:
  /auth/kakao/callback:
    get:
      description: 카카오 code를 받아 로그인 또는 회원가입 필요 상태를 반환합니다.
      parameters:
      - description: 인가 코드
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 카카오 OAuth 콜백 처리
      tags:
      - Auth
  /auth/kakao/login-url:
    get:
      description: 프론트에서 카카오 인증 화면으로 이동할 URL을 조회합니다.
      parameters:
      - default: http://localhost:8080/api/v1/auth/kakao/callback
        description: 로그인 완료 후 리다이렉트 URI
        in: query
        name: redirectUri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 카카오 OAuth 로그인 URL 조회
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: 로그아웃 요청
        in: body
        name: request
        schema:
          $ref: '#/definitions/handlers.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      - RefreshTokenAuth: []
      summary: 로그아웃
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: 리프레시 토큰 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      - RefreshTokenAuth: []
      summary: 액세스 토큰 재발급
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: signupToken 기반으로 회원가입 후 로그인 토큰을 발급합니다.
      parameters:
      - description: 회원가입 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 회원가입
      tags:
      - Auth
  /auth/session:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - AccessTokenAuth: []
      summary: 현재 인증 세션 조회
      tags:
      - Auth
...
securityDefinitions:
  AccessTokenAuth:
    description: 'Bearer access token. Example: Bearer eyJ...'
    in: header
    name: Authorization
    type: apiKey
  RefreshTokenAuth:
    description: 'Bearer refresh token. Example: Bearer eyJ...'
    in: header
    name: X-Refresh-Token
    type: apiKey
swagger: "2.0"